Reorganize to enable importing elsewhere
